refactor: move authentication middleware test to a separed file
